requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
google-auth==1.27.1
requests==2.25.1
firebase-admin==4.5.3
firebase==3.0.1
Flask-Session==0.3.2